requirements.txt
affine==2.4.0
asttokens==3.0.0
attrs==25.3.0
branca==0.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
click-plugins==1.1.1
cligj==0.7.2
comm==0.2.2
contextily==1.6.2
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
executing==2.2.0
folium==0.19.6
fonttools==4.58.0
geographiclib==2.0
geojson==3.2.0
geopandas==1.0.1
geopy==2.4.1
greenlet==3.2.2
idna==3.10
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.8
mapclassify==2.9.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mercantile==1.2.1
nest-asyncio==1.6.0
networkx==3.5
numpy==2.2.6
packaging==25.0
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
prompt_toolkit==3.0.51
psutil==7.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
pyogrio==0.11.0
pyparsing==3.2.3
pyproj==3.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
pyzmq==26.4.0
rasterio==1.4.3
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
shapely==2.1.1
six==1.17.0
SQLAlchemy==2.0.41
stack-data==0.6.3
threadpoolctl==3.6.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
wcwidth==0.2.13
xyzservices==2025.4.0
